logic set up to read data specific to user
