+++ app/Models/home.php
@@ -5,7 +5,9 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
-class Event extends Model
+class home extends Model
 {
     use HasFactory;
+    protected $table = 'homes';
+    public $primaryKey = 'email';
 }
